requirements.txt
# This is the output of pip freeze in the environment used for the experiments, with a couple of tweaks so
# that you can install them without dependency problems.
# Note that not all of these modules are required.
# Just make sure to install 'nvidia-tensorflow[horovod]' and then add the remaining needed libraries as they raise an error.
# I want to make your life easier, so just use pip install -r requirements.txt and you'll have the same env I had.

# Also, make sure your CUDA dependencies are in your PATH.

absl-py==0.11.0
alabaster==0.7.12
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
astropy==4.0
astunparse==1.6.3
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.8.2
bitarray==1.2.0
bkcharts==0.2
bleach==3.1.0
bokeh==1.4.0
boto==2.49.0
Bottleneck==1.3.1
cached-property==1.5.2
cachetools==4.2.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloudpickle==1.6.0
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==2.8
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.9.2
dataclasses
decorator==4.4.1
defusedxml==0.6.0
distributed==2.9.3
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
ffmpeg==1.4
Flask==1.1.1
flatbuffers==1.12
fsspec==0.6.2
gast==0.2.2
gevent==1.4.0
gmpy2==2.0.8
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.34.1
h5py==3.1.0
hdf5storage==0.1.15
HeapDict==1.0.1
html5lib==1.0.1
hypothesis==4.54.2
idna==2.8
imageio==2.6.1
imagesize==1.2.0
importlib-metadata==3.4.0
imutils==0.5.3
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.2
jeepney==0.4.2
Jinja2==2.10.3
joblib==0.14.1
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.5
jupyterlab-server==1.0.6
kaggle==1.5.10
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==20.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
locket==0.2.0
lxml==4.4.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
more-itertools==8.0.2
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
nose==1.3.7
notebook==6.0.2
numba==0.47.0
numexpr==2.7.0
numpy
nvidia-tensorflow[horovod]
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.2.30
openpyxl==3.0.2
opt-einsum==3.3.0
packaging==20.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
partd==1.1.0
path==13.1.0
pathlib2==2.3.5
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.2
protobuf==3.14.0
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pyfakewebcam==0.1.0
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pyodbc==4.0.27
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyqtgraph==0.10.0
pyrsistent==0.15.6
PySocks==1.7.1
pytest==5.3.2
pytest-arraydiff==0.3
pytest-astropy==0.7.0
pytest-astropy-header==0.1.1
pytest-doctestplus==0.5.0
pytest-openfiles==0.4.0
pytest-remotedata==0.3.2
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==18.1.0
QtAwesome==0.6.0
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
requests-oauthlib==1.3.0
rope==0.14.0
rsa==4.7
ruamel-yaml==0.15.87
scikit-image==0.15.0
scikit-learn==0.22.1
scipy
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
shiboken6==6.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
snowballstemmer==2.0.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.9.5
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.2
SQLAlchemy==1.3.12
statsmodels==0.10.1
sympy==1.5.1
tables==3.6.1
tb-nightly==2.5.0a20210115
tblib==1.6.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-addons==0.12.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tf-estimator-nightly==2.5.0.dev2021011501
toolz==0.10.0
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
typed-ast==1.4.0
typeguard==2.7.1
typing-extensions==3.7.4.3
unicodecsv==0.14.1
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wurlitzer==2.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
zict==1.0.0
zipp==3.4.0
